desktop/src-tauri/resources/copaw/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sse-starlette
